requirements.txt
fastapi==0.112.1
uvicorn[standard]==0.30.6
pydantic==2.8.2
requests==2.32.3
weaviate-client==4.6.4
tiktoken==0.7.0
python-dotenv==1.0.1
comet-ml==3.43.0
httpx==0.27.2
tenacity==8.5.0
PyJWT==2.9.0
pandas==2.2.2
jinja2==3.1.4
pytest==8.3.2